% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movebank.R
\name{movebank}
\alias{movebank}
\title{movebank}
\usage{
movebank(table = "study", ..., user = Sys.getenv("MOVEBANK_USER"),
  password = Sys.getenv("MOVEBANK_PW"))
}
\arguments{
\item{table}{Name of the table, see details}

\item{...}{additional fields to filter on as named list, depends on table selecte}

\item{user}{username. Will be read from Sys.getenv("MOVEBANK_USER") if not supplied, register at https://www.movebank.org/}

\item{password}{user password. Will be read from Sys.getenv("MOVEBANK_PW") if not supplied, register at https://www.movebank.org/}
}
\value{
Returns the requested table (read by read_csv), possibly filtered by additional search fields
}
\description{
Read data from movebank
}
\examples{
\dontrun{

## Not run, requires password set
movebank("study")

}
}
